renamed package `fitmodel` to `math`
